First Commit
We’ve been messing around with some code and hopefully have finally
figured out the majority of the physics. Old assets from previous tests
and attempts have been left in the project. Tilt script not quite right
at this point, but that’s for non-Arduino users, used for testing
purposes.